[refactor/#5] Toast 확장함수 적용
