+++ app/src/main/java/com/example/soptseminar1/presentation/profile/ProfileFragment.kt
@@ -3,17 +3,17 @@ package com.example.soptseminar1.presentation.profile
 import android.content.Intent
 import android.os.Bundle
 import android.view.View
-import android.widget.Toast
 import com.bumptech.glide.Glide
 import com.example.soptseminar1.R
 import com.example.soptseminar1.data.api.GithubApiCreator
 import com.example.soptseminar1.data.model.response.ResponseGithubUserInformation
 import com.example.soptseminar1.databinding.FragmentProfileBinding
-import com.example.soptseminar1.enqueueUtil
+import com.example.soptseminar1.util.enqueueUtil
 import com.example.soptseminar1.presentation.profile.follower.FollowerFragment
 import com.example.soptseminar1.presentation.profile.logout.LogoutActivity
 import com.example.soptseminar1.presentation.profile.repository.RepositoryFragment
 import com.example.soptseminar1.util.BaseFragment
+import com.example.soptseminar1.util.showToast
 import retrofit2.Call
 
 class ProfileFragment : BaseFragment<FragmentProfileBinding>(FragmentProfileBinding::inflate) {
@@ -50,7 +50,7 @@ class ProfileFragment : BaseFragment<FragmentProfileBinding>(FragmentProfileBind
                 }
             },
             onError = {
-                Toast.makeText(requireContext(), "불러오기에 실패했습니다.", Toast.LENGTH_SHORT).show()
+                requireContext().showToast("불러오기에 실패했습니다.")
             }
         )
     }
