+++ app/src/main/java/com/example/soptseminar1/presentation/profile/logout/LogoutActivity.kt
@@ -2,10 +2,9 @@ package com.example.soptseminar1.presentation.profile.logout
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.widget.Toast
-import com.example.soptseminar1.R
 import com.example.soptseminar1.data.api.SOPTSharedPreferences
 import com.example.soptseminar1.databinding.ActivityLogoutBinding
+import com.example.soptseminar1.util.showToast
 
 class LogoutActivity : AppCompatActivity() {
 
@@ -22,7 +21,7 @@ class LogoutActivity : AppCompatActivity() {
     private fun logoutBtnClink(){
         binding.ivLogout.setOnClickListener {
             SOPTSharedPreferences.setLogout(this)
-            Toast.makeText(this, "자동 로그인이 해제되었습니다.", Toast.LENGTH_SHORT).show()
+            showToast("자동 로그인이 해제되었습니다.")
         }
     }
 }
