+++ app/src/main/java/com/example/soptseminar1/presentation/profile/repository/RepositoryFragment.kt
@@ -2,13 +2,13 @@ package com.example.soptseminar1.presentation.profile.repository
 
 import android.os.Bundle
 import android.view.View
-import android.widget.Toast
 import com.example.soptseminar1.data.api.GithubApiCreator
 import com.example.soptseminar1.data.model.response.ResponseGithubUserRepo
 import com.example.soptseminar1.databinding.FragmentRepositoryBinding
-import com.example.soptseminar1.enqueueUtil
+import com.example.soptseminar1.util.enqueueUtil
 import com.example.soptseminar1.util.MyItemDecoration
 import com.example.soptseminar1.util.BaseFragment
+import com.example.soptseminar1.util.showToast
 import retrofit2.Call
 
 class RepositoryFragment : BaseFragment<FragmentRepositoryBinding>(FragmentRepositoryBinding::inflate) {
@@ -37,7 +37,7 @@ class RepositoryFragment : BaseFragment<FragmentRepositoryBinding>(FragmentRepos
                 }
             },
             onError = {
-                Toast.makeText(requireContext(), "불러오기에 실패했습니다.", Toast.LENGTH_SHORT).show()
+                requireContext().showToast("불러오기에 실패했습니다.")
             }
         )
     }
