+++ app/src/main/java/com/example/soptseminar1/presentation/signup/SignUpActivity.kt
@@ -3,13 +3,13 @@ package com.example.soptseminar1.presentation.signup
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.widget.Toast
 import com.example.soptseminar1.presentation.signin.SignInActivity
 import com.example.soptseminar1.data.api.ServiceCreator
 import com.example.soptseminar1.data.model.request.RequestSignUp
 import com.example.soptseminar1.data.model.response.ResponseSignUp
 import com.example.soptseminar1.databinding.ActivitySignUpBinding
-import com.example.soptseminar1.enqueueUtil
+import com.example.soptseminar1.util.enqueueUtil
+import com.example.soptseminar1.util.showToast
 import retrofit2.Call
 
 class SignUpActivity : AppCompatActivity() {
@@ -48,8 +48,8 @@ class SignUpActivity : AppCompatActivity() {
             },
             onError = {
                 when(it){
-                    409 -> Toast.makeText(this@SignUpActivity, "이미 존재하는 유저입니다.", Toast.LENGTH_SHORT).show()
-                    else -> Toast.makeText(this@SignUpActivity, "회원가입에 실패하였습니다.", Toast.LENGTH_SHORT).show()
+                    409 -> showToast("이미 존재하는 유저입니다.")
+                    else -> showToast("회원가입에 실패하였습니다.")
                 }
             }
         )
